net: Move NetEmbedderMsg to the net crate.
