+++ components/net/request_interceptor.rs
@@ -3,13 +3,14 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 use content_security_policy::Destination;
-use embedder_traits::{EmbedderProxy2, NetEmbedderMsg, WebResourceRequest, WebResourceResponseMsg};
+use embedder_traits::{EmbedderProxy2, WebResourceRequest, WebResourceResponseMsg};
 use log::error;
 use net_traits::NetworkError;
 use net_traits::http_status::HttpStatus;
 use net_traits::request::Request;
 use net_traits::response::{Response, ResponseBody};
 
+use crate::embedder::NetEmbedderMsg;
 use crate::fetch::methods::FetchContext;
 
 #[derive(Clone)]
